end is funqcion cga, return, errorPatg, uploaddir, root

+++ ParsingConfigFile.hpp
@@ -44,7 +44,7 @@ private:
 	size_t	checkUploadDir(size_t pos_start, size_t pos_end, bool controlFlag);
 
 	size_t	checkErrorPage(size_t pos_start, size_t pos_end, bool controlFlag);
-	int		statusCodes(std::string number);
+	size_t	statusCodes(std::string number);
 
 	size_t	checkReturn(size_t pos_start, size_t pos_end, bool controlFlag);
